dummy open shop (its really just jobshop)
